+++ CHANGES.md
@@ -1,6 +1,7 @@
 # 0.5.4 / 2021-02-22
 
   * Scoring by Orme (2009) added to docs
+  * Bradley-Terry-Luce (BTL) Model added
 
 # 0.5.2 / 2021-02-22